requirements.txt
wheel
python-dotenv
langchain
pypdf
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic==1.10.8
rapidocr-onnxruntime
openai
tiktoken
faiss-gpu
